+++ src/reader.h
@@ -29,7 +29,7 @@ typedef struct
 	FILE           *file;
 	int             eof;
 	int             seekable;
-	long            file_size;
+	size_t          file_size;
 
 	int             sockfd;
 
@@ -43,14 +43,14 @@ typedef struct
 	pthread_mutex_t mutex;
 	pthread_t       thread;
 
-	unsigned long   stream_pos;
+	size_t          stream_pos;
 
 	int             is_ready;
 } Reader;
 
 /* Opens a local file or HTTP URL for reading */
-int     reader_set_cache_size_kb(size_t size, size_t prebuffer_size);
-int     reader_get_cache_fill(Reader *r);
+size_t  reader_set_cache_size_kb(size_t size, size_t prebuffer_size);
+size_t  reader_get_cache_fill(Reader *r);
 Reader *reader_open(const char *url);
 int     reader_close(Reader *r);
 int     reader_is_ready(Reader *r);
@@ -62,10 +62,10 @@ size_t  reader_get_number_of_bytes_in_buffer(Reader *r);
 /* Resets the stream to the beginning (if possible), returns 1 on success, 0 otherwise */
 int     reader_reset_stream(Reader *r);
 int     reader_is_seekable(Reader *r);
-int     reader_seek_whence(Reader *r, long byte_offset, int whence);
-int     reader_seek(Reader *r, long byte_offset);
-long    reader_get_file_size(Reader *r);
-unsigned long reader_get_stream_position(Reader *r);
+int     reader_seek_whence(Reader *r, ssize_t byte_offset, int whence);
+int     reader_seek(Reader *r, ssize_t byte_offset);
+size_t  reader_get_file_size(Reader *r);
+size_t  reader_get_stream_position(Reader *r);
 /* Sets number of bytes in buffer to 0 */
 void    reader_clear_buffer(Reader *r);
 #endif
